Add linked action execution in GRPC tag service
A new RPC method "ExecuteLinkedAction" is added to the TagService which allows execution of actions linked with files or folders. This change enables clients to execute specified actions on either files or folders depending upon the request received. The system processing is performed through the newly implemented 'ExecuteLinkedAction' method in the TagService.cs file and the corresponding request and reply messages were defined in the 'TagService.proto' file.