Added trash functionality for file deletion
Added trash directory to the CommonStorageOptions and implemented a 'MoveFileToTrash' function in CommonStorage. This allows files to be moved to trash rather than being deleted outright, hence providing a way to recover accidentally deleted files. This change affects the DeleteFile command, which now utilizes this new function. Appropriate changes were also made in FileActionsService to complete the GRPC endpoint for deleting files.